Squashed commit of the following:
commit 324786ced499b0aae48edc6006c0a68592106db2
Author: An Jun <limanjun99@gmail.com>
Date:   Sun Jan 14 22:15:17 2024 +0800

    feat: add killer move heuristic

commit 057ccebcc26d9ba128cf44f12461fe9694f48459
Author: An Jun <limanjun99@gmail.com>
Date:   Sun Jan 14 22:09:25 2024 +0800

    fix: en passant wrongly treated as not a capture

commit 79fc24a7f83a734640af06053484b43565139325
Author: An Jun <limanjun99@gmail.com>
Date:   Sun Jan 14 19:04:26 2024 +0800

    chore: add captured field to Move

+++ src/chess/include/chess/piece.h
@@ -2,7 +2,7 @@
 
 #include <cstdint>
 
-enum class Piece : int32_t {
+enum class Piece : int8_t {
   Bishop = 0,
   King = 1,
   Knight = 2,
